remove floating point errors to ensure accurate md5sum reproducibility of chic score

+++ src/mixer/utils/shuffle/scoring/ShuffleScore.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2011-2021 Rice University, Baylor College of Medicine, Aiden Lab
+ * Copyright (c) 2011-2022 Rice University, Baylor College of Medicine, Aiden Lab
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -35,17 +35,17 @@ public abstract class ShuffleScore {
         this.cBounds = cBounds;
     }
 
-    public double score(boolean isBaseline) {
+    public float score(boolean isBaseline) {
         if (isBaseline) {
             return baselineScore();
         }
         return score(rBounds, cBounds);
     }
 
-    private double baselineScore() {
+    private float baselineScore() {
         return score(new Integer[]{0, matrix.length},
                 new Integer[]{0, matrix[0].length});
     }
 
-    protected abstract double score(Integer[] rBounds, Integer[] cBounds);
+    protected abstract float score(Integer[] rBounds, Integer[] cBounds);
 }
